moved individual dashboard to src directory
