+++ src/components/Individual/actions.js
