+++ src/components/Individual/edit.js
@@ -3,9 +3,9 @@ import Image from 'next/image'
 import React, { useRef, useState } from 'react'
 import { BsCamera } from 'react-icons/bs'
 import localFont from 'next/font/local'
-import profileimage from "../../public/images/individual/profileimage.png";
+import profileimage from "../../../public/images/individual/profileimage.png";
 
-const myFont = localFont ({ src: '../../public/fonts/IBMPlexMono-SemiBold.woff2' })
+const myFont = localFont ({ src: '../../../public/fonts/IBMPlexMono-SemiBold.woff2' })
 const Edit = () => {
     const user={
         DOB:"",
